PVWS replacement: add base pvaPy files
For this application, accessing all record fields is essential (not only
VAL), especially if using OPIs for motors, for example. PVWS does not
support specific field subscription as of now, so we cannot use it on
the long run. This commit is the first step towards the refactor of the
WebSocket layer.

I decided to use pvapy: it covers everything and is as efficient as it
gets.
Main ideas of PVWS are kept (json message format, b64 encoding, etc.)